switch to doubles

+++ helper.ino
@@ -1,11 +1,11 @@
-const void euler_to_quat(float roll, float pitch, float yaw, double *q)
+const void euler_to_quat(double roll, double pitch, double yaw, double *q)
 {
-    float cr = cos(roll * 0.5);
-    float sr = sin(roll * 0.5);
-    float cp = cos(pitch * 0.5);
-    float sp = sin(pitch * 0.5);
-    float cy = cos(yaw * 0.5);
-    float sy = sin(yaw * 0.5);
+    double cr = cos(roll * 0.5);
+    double sr = sin(roll * 0.5);
+    double cp = cos(pitch * 0.5);
+    double sp = sin(pitch * 0.5);
+    double cy = cos(yaw * 0.5);
+    double sy = sin(yaw * 0.5);
 
     q[0] = cr * cp * cy + sr * sp * sy;
     q[1] = sr * cp * cy - cr * sp * sy;
@@ -14,14 +14,14 @@ const void euler_to_quat(float roll, float pitch, float yaw, double *q)
 }
 // inputs actual velocity for left and right wheel, from odrive data
 // outputs linear vel in x dir
-float generateLinearVel(float lwvel, float rwvel)
+double generateLinearVel(double lwvel, double rwvel)
 {
     return GEARRATIO * (WHEELRAD / 2) * (wheelc1 * lwvel + wheelc2 * rwvel);
 }
 
 // inputs actual velocity for l and r wheel, from odrive data
 // outputs angular vel in z dir
-float generateAngularVel(float lwvel, float rwvel)
+double generateAngularVel(double lwvel, double rwvel)
 {
     return GEARRATIO * (WHEELRAD / 2) * ((wheelc1 * lwvel - wheelc2 * rwvel) / WHEELSEP);
 }
